requirements.txt
smbus2
pynmea2
pyserial
raspy-qmc5883l
pandas